[feat/#86]: 컴포넌트 분리 및 코드 간결화
